==Fixed== Issue 1: When starting up, if FB2K is set to remember what it was doing when the user last closed, rich presence would always announce playing even if paused. Now announces the correct state when starting up. Changes are in foo_drpc::on_playback_new_track.
==Changed==
Plugin.h now has a placeholder for APPLICATION_ID. It now checks secret.h for this value. This is being done so we don't have to be extra careful not to submit our application id (which is a secret) when committing to git. README.md is updated to reflect as much.
Since we want the file to persist, people using this must run the following command to ignore local changes to the file:
git update-index --assume-unchanged foo_drpc/secret.h